requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
pyyaml>=6.0.1
python-multipart>=0.0.7
chromadb>=0.4.22
sentence-transformers>=2.3.0
aiofiles>=23.2.1
